+++ spec/celluloid/io/actor_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
-describe Celluloid::IO::Actor do
+describe Celluloid::IO do
   it_behaves_like "a Celluloid Actor", Celluloid::IO
 end